feat(rules): an emplacement has a position, not a slot (#116)
* feat(rules): an emplacement has a position, not a slot

Closes PLC-01 (#29). Representation only — the legality test is a stub that accepts exactly
the anchor's authored slot positions, so nothing about behaviour may change and nothing did.

    parity ok — 1440 runs identical (gdscript vs python)
    (digest 2efa855e11f7, freshly verified — not from cache)

PILLAR 2 OF THE PRD IS "build where you want; fixed slots are the single most limiting thing
in the current design." This is the representation half of that. It is also the fix for what
the owner actually complained about: LF-176 was "most turrets can barely reach anything",
and decision 074's range pass proved the residual is SLOT GEOMETRY, not range values —
sweeping pulse-turret from 3.2 to 7.0, more than double what shipped, leaves p10 pinned at
literal 0.0% while the median climbs sixteen points. No tuning reaches a pad that is simply
too far from the lane. Only moving the pad does.

THE SEAM. `Placed.slot: tuple[int,int]` becomes `x`/`y` float64 in both engines, and every
legality question goes through ONE function, mirrored:

    # sim/engine.py
    def _is_placeable(self, x: float, y: float) -> bool:
        return (x, y) in self.a.slots

    # scripts/anchor_sim.gd
    func _is_placeable(x: float, y: float) -> bool:
        for s in anchor.get("slots", []):
            if float(s[0]) == x and float(s[1]) == y:
                return true
        return false

Python's `in` over a tuple of floats does the identical elementwise equality the GDScript
loop does explicitly. PLC-02 replaces exactly these two bodies and nothing else. Doing
representation and rules in one commit would have meant no way to tell which broke parity —
that is the issue's own reasoning and it held.

`Placed` gains a read-only `slot` property returning `(x, y)` so `sim/coverage.py` — out of
scope and deliberately untouched — keeps reading real data. Proved with its own selftest:
5/5, including Sim/InstrumentedSim outcome parity.

AND THE PART WORTH MORE THAN THE DIFF: four places still assume integer slot coordinates and
will bite PLC-02, found by looking rather than by waiting. All are presentation; none is in
the rules.

  1. combat_fx.gd's `_beams` is keyed by Vector2i(int(x), int(y)). Two emplacements in the
     SAME integer cell — (4.3, 7.0) and (4.7, 7.0) — collide, and _find_placed() returns
     whichever it hits first, so a sustained beam can visually attach to the wrong gun. This
     one is load-bearing-wrong, not merely stale.
  2. AnchorSim.available_slots() enumerates anchor["slots"] and returns Vector2i, so it is
     tied to the fixed-slot concept rather than to _is_placeable(). The moment PLC-02 accepts
     a position outside that array this function will not reflect it AT ALL, and every
     consumer needs rework, not a signature change.
  3. The whole click/select UI is Vector2i-typed, and `hovered_slot = Vector2i(roundi(t.x),
     roundi(t.y))` rounds the mouse before anything else sees it. There is currently no way
     for a player to build at a non-integer position regardless of what the rules allow —
     which is exactly how PLC-02 could ship and change nothing visible in play.
  4. main.gd's scenario "build" verb truncates with int(), so a fractional build position in
     a scenario JSON would be silently floored.

Cosmetic only: fx_additive.gd's pulse-desync hash also rounds; a collision there just means
two supports pulse in phase.

Confirmed already float-native and ready for PLC-02 to swap the predicate body alone:
_is_placeable, _occupied, _slot_priority, _covered_by, _select_target, and
IsoScript.tile_to_screen. One int() cast that is CORRECT and not a landmine —
anchor_view.gd's _height_at(int(x), int(y)) — that is the intended world-to-tile truncation
matching sim/content.py's own Anchor.height_at().

VERIFICATION, all four bars the issue set:
  - rules parity: 1440 runs identical, freshly verified, not from cache
  - 24 anchors x 3 difficulties: 0 field differences against a git-worktree baseline built
    from HEAD, compared field-for-field rather than eyeballed
  - tier 2: 25 passed, 0 failed; `safe operations` and `rules autoloads` both ok
  - tools/shot.py anchor-01 --build pulse-turret --build scan-relay renders correctly with
    towers on their slots and the reactor bus right

tools/analysis_lf014.py and tools/bench_tick.py construct Placed() directly and are updated.
tools/parity_costs.json is test_parity.py's own telemetry write on a full run.


* docs(backlog): PLC-02's landmines, and a shipped feature the owner never heard about

Five items from this branch's work and from the owner's own play question.

LF-197 — the four remaining integer-slot assumptions PLC-01 went looking for. Full detail
in the feature commit; the one that matters most is that the click/select UI rounds the
mouse to Vector2i before anything else sees it, so PLC-02 could ship green and change
NOTHING visible in play. Handed to #30 as a comment rather than left in a commit body.

LF-198 — tools/analysis_overdraw.py and analysis_shield_wall.py are dead AND now broken.
They were already non-runnable before this change: both hardcode
sys.path.insert(0, '/Users/briangerhards/dev/defend-claude'), an old project name on an old
host. Recommend deleting rather than repairing. analysis_lf014.py was the same family but
IS repaired, because it was reachable.

LF-199 — LF-123 and LF-126 name AnchorSim.free_slots, a field that no longer exists. Both
are stale WORDING, not stale findings. Filed rather than silently edited, because LF-195
just established that quietly-wrong items cost a dispatch.

LF-177 CLOSED — the game speed-up the owner asked for is not deferred, it SHIPPED. Backtick
cycles 1x/2x/3x from data/tuning.json's authored speeds, gamepad button 7 too, and the HUD
prints the current value top-left. Implemented exactly the safe way LF-177 itself demanded:
_accum += delta * speed runs the SAME fixed tick more times per frame rather than scaling
dt, which is why parity is unaffected.

LF-196 — and that is the actual finding. The owner asked about pacing again today, at 1x,
because nobody told them. A shipped player-facing control the owner does not know exists is
worth nothing, and the backlog said "deferred" long enough to be quoted back as outstanding.
Two things follow: audit every player-facing control against what the owner has actually
been told (speed is unlikely to be the only one — three abilities on 1/2/3 and an
on/off toggle on F are also undocumented anywhere they would see), and treat "shipped but
unannounced" as a distinct state from "open", because closing something silently is what
let this happen. Worse than LF-195: this one cost the owner a play session.


* docs(chronicle): a position, not a slot

Journal entry for PLC-01. Pillar 2 of the PRD starts moving: an emplacement carries x/y
float64 instead of an integer slot, in both engines, with parity 1440 runs identical and
freshly verified rather than served from cache — a cached pass would have proved nothing
about a representation change.

The entry leads on the part worth more than the diff: the agent was asked whether anything
still assumed integer slot coordinates, and found four. The one that matters is that the
click/select UI rounds the mouse to Vector2i before anything else sees it, so PLC-02 could
ship green and change nothing visible in play.

chronicle.py --check green, including the markup audit that landed two PRs ago. Previous
entry changed by its nav link only.


---------

+++ sim/content.py
@@ -145,7 +145,12 @@ class Anchor:
     lives: int
     grid: tuple[int, int]
     lanes: tuple[Lane, ...]
-    slots: tuple[tuple[int, int], ...]
+    # PLC-01: float64, not int — Placed.x/y (sim/engine.py) are continuous positions and
+    # the legality stub (Sim._is_placeable()) compares a build position against this tuple
+    # directly, so both sides of the comparison must be the same type. `float(x)` on an
+    # authored integer is exact (every one of today's 24 anchors), matching the loader's
+    # existing idiom for `waypoints` above.
+    slots: tuple[tuple[float, float], ...]
     waves: tuple[Wave, ...]
     tutorial: bool = False
     # Act III: MW the bus loses at the start of every wave after the first. The reactor
@@ -383,7 +388,9 @@ def load_anchor(anchor_id: str) -> Anchor:
         # missing `slots` key the same way (`anchor.get("slots", [])`); this loader was the
         # one place still requiring it, which would crash rather than degrade the moment a
         # free-placement anchor existed to load.
-        slots=tuple((int(x), int(y)) for x, y in doc.get("slots", [])),
+        #
+        # PLC-01: `float(x)`, not `int(x)` — see the field's own comment on Anchor.slots.
+        slots=tuple((float(x), float(y)) for x, y in doc.get("slots", [])),
         waves=tuple(
             Wave(
                 lead_in=w.get("lead_in", 20.0),
